Run lint on it
Lesson learnt, do not make silly changes on github edit but not test it, and also run pylint on your code.

OK this code got 0/10 pylint marks because of long lines and not enough indentation and use of globals, but I have a difference of opinion with python about line length when modern lcd's will do 120 chars on a line side by side easy.  And I think it is a weakness of python to not have block delimiters and instead rely purely on indentation as block delimiters.  But, I really like python as well.